+++ src/Sulu/Bundle/MediaBundle/Controller/MediaStreamController.php
@@ -20,9 +20,9 @@ use Sulu\Bundle\MediaBundle\Media\FormatManager\FormatManagerInterface;
 use Sulu\Bundle\MediaBundle\Media\Manager\MediaManagerInterface;
 use Sulu\Bundle\MediaBundle\Media\Storage\StorageInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\HttpFoundation\BinaryFileResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\ResponseHeaderBag;
-use Symfony\Component\HttpFoundation\StreamedResponse;
 
 class MediaStreamController extends Controller
 {
@@ -67,7 +67,7 @@ class MediaStreamController extends Controller
      * @param Request $request
      * @param int $id
      *
-     * @return StreamedResponse
+     * @return BinaryFileResponse
      */
     public function downloadAction(Request $request, $id)
     {
@@ -103,7 +103,7 @@ class MediaStreamController extends Controller
      * @param string $locale
      * @param string $dispositionType
      *
-     * @return StreamedResponse
+     * @return BinaryFileResponse
      */
     protected function getFileResponse(
         $fileVersion,
@@ -120,16 +120,7 @@ class MediaStreamController extends Controller
 
         $path = $this->getStorage()->load($fileName, $version, $storageOptions);
 
-        $response = new StreamedResponse(function () use ($path) {
-            flush(); // send headers
-            $handle = fopen($path, 'r');
-            while (!feof($handle)) {
-                $buffer = fread($handle, 1024);
-                echo $buffer;
-                flush(); // buffered output
-            }
-            fclose($handle);
-        });
+        $response = new BinaryFileResponse($path);
 
         $pathInfo = pathinfo($fileName);